AMB-1745 refactor
