+++ tests/test_crud_immunisation_api.py
@@ -2,43 +2,9 @@ import copy
 import uuid
 
 import pytest
-import requests
 
 from .example_loader import load_example
-
-
-class ImmunisationApi:
-
-    def __init__(self, url, token):
-        self.url = url
-        self.token = token
-        self.headers = {
-            "Authorization": self.token,
-            "Content-Type": "application/fhir+json",
-            "Accept": "application/fhir+json",
-        }
-
-    def get_immunization_by_id(self, event_id):
-        return requests.get(f"{self.url}/event/{event_id}", headers=self._update_headers())
-
-    def create_immunization(self, imms):
-        return requests.post(f"{self.url}/event", headers=self._update_headers(), json=imms)
-
-    def delete_immunization(self, imms_id):
-        return requests.delete(f"{self.url}/event/{imms_id}", headers=self._update_headers())
-
-    def search_immunizations(self, nhs_number, disease_type):
-        return requests.get(f"{self.url}/event?nhsNumber={nhs_number}&diseaseType={disease_type}",
-                            headers=self._update_headers())
-
-    def _update_headers(self, headers=None):
-        if headers is None:
-            headers = {}
-        updated = {**self.headers, **{
-            "X-Correlation-ID": str(uuid.uuid4()),
-            "X-Request-ID": str(uuid.uuid4()),
-        }}
-        return {**updated, **headers}
+from .immunization_api import ImmunisationApi
 
 
 def create_an_imms_obj(imms_id: str = str(uuid.uuid4())) -> dict:
